% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_histograms.R
\name{create_sample_histograms}
\alias{create_sample_histograms}
\title{Create a sample histogram from a population}
\usage{
create_sample_histograms(pop, samples, var_name, sample_size)
}
\arguments{
\item{pop}{the virtual population as a tibble}

\item{samples}{the samples as a tibble}

\item{var_name}{the name of the column for the variable that is being generated}

\item{sample_size}{a vector of the sample sizes}
}
\value{
a list of the sample histogram plots
}
\description{
This function creates a grid of sample distributions from a population for different sample sizes.
Author: Holly Williams
}
\examples{
pop <- generate_virtual_pop(100, height, rnorm, 0, 1)
samples <- draw_samples(pop, 3, c(1, 10))
create_sample_histograms(pop, samples, height, c(1, 10))
}
