% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_samples.R
\name{draw_samples}
\alias{draw_samples}
\title{Generates samples of different sizes}
\usage{
draw_samples(pop, reps, n_s)
}
\arguments{
\item{pop}{the virtual population as a tibble}

\item{reps}{the number of replication for each sample size as an integer}

\item{n_s}{the sample size for each one of the samples as an array}
}
\value{
a two column tibble containing the sample numbers and sample values
}
\description{
This function draws samples of various sizes from a population.
Author: Lise Braaten
}
\examples{
pop <- generate_virtual_pop(100, "Variable", rnorm, 0, 1)
samples <- draw_samples(pop, 3, c(1, 10))
}
