% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_virtual_pop.R
\name{generate_virtual_pop}
\alias{generate_virtual_pop}
\title{Create a virtual population}
\usage{
generate_virtual_pop(N, var_name, dist, ...)
}
\arguments{
\item{N}{The numbers of samples}

\item{var_name}{the variable name that we need to create}

\item{dist}{the function that we are using}

\item{...}{for the dist function}
}
\value{
a tibble that stores the virtual population
}
\description{
This function creates a population from a distribution.
}
\examples{
generate_virtual_pop(100, "height", rnorm, 0, 1)
generate_virtual_pop(100, "n_accidents_per_hr", rpois, 2)

}
