% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_hist.R
\name{plot_sample_hist}
\alias{plot_sample_hist}
\title{Create a sample histogram from a population}
\usage{
plot_sample_hist(pop, samples, var_name, n_s)
}
\arguments{
\item{pop}{the virtual population as a tibble}

\item{samples}{the samples as a tibble}

\item{var_name}{the name of the variable/parameter of interest that is being generated}

\item{n_s}{a vector of the sample sizes}
}
\value{
a grid of the sample distributions
}
\description{
This function creates a grid of sample distributions from a population.
Author: Holly Williams
}
\examples{
pop <- generate_virtual_pop(100, "Variable", rnorm, 0, 1)
samples <- draw_samples(pop, 3, c(1, 10))
plot_sample_hist(pop, samples, var_name, c(1, 10))
}
