% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary.R
\name{stat_summary}
\alias{stat_summary}
\title{Create a stats summary}
\usage{
stat_summary(population, samples, parameter)
}
\arguments{
\item{population}{The virtual population}

\item{samples}{The drawed samples}

\item{parameter}{A string vector or a single string of parameter(s) of interest}
}
\value{
a tibble summary stats
}
\description{
This function creates a summary stats for population, samples and parameter(s) of interest.
}
\examples{
pop <- generate_virtual_pop(100000, "Height", rexp, 5)
samples <- draw_samples(pop, 1, c(1))
stat_summary(pop, samples, c('mean', 'sd'))

}
