% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_hist.R
\name{plot_sample_hist}
\alias{plot_sample_hist}
\title{Create a sample histogram from a population}
\usage{
plot_sample_hist(pop, samples, var_name, sample_size)
}
\arguments{
\item{pop}{the virtual population as a tibble}

\item{samples}{the samples as a tibble}

\item{var_name}{the name of the column for the variable that is being generated}

\item{sample_size}{a vector of the sample sizes}
}
\value{
a list of the sample histogram plots
}
\description{
This function creates a grid of sample distributions from a population for different sample sizes.
The plotted sample is the first replication from the samples (rep == 1). The other replicates
are used for plotting the sampling distribution (applying the \code{plot_sampling_hist} function).
Author: Holly Williams
}
\examples{
pop <- generate_virtual_pop(100, height, rnorm, 0, 1)
samples <- draw_samples(pop, 3, c(1, 10))
plot_sample_hist(pop, samples, height, c(1, 10))
}
