% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampling_hist.R
\name{plot_sampling_hist}
\alias{plot_sampling_hist}
\title{Create a sampling distribution histogram from samples}
\usage{
plot_sampling_hist(samples, var_name, n_s, reps)
}
\arguments{
\item{samples}{the samples as a tibble}

\item{var_name}{the name of the variable/parameter of interest that is being generated}

\item{n_s}{a vector of the sample sizes}

\item{reps}{the number of replication for each sample size as an integer}
}
\value{
a list of the sampling distributions
}
\description{
This function creates a list of sampling distribution histogram of the mean of different sample sizes drawn from a population
Author: Yue(Alex) Jiang
}
\examples{
pop <- generate_virtual_pop(100, "Variable", rnorm, 0, 1)
samples <- draw_samples(pop, 3, c(1, 10))
plot_sampling_hist(samples, Variable, c(1, 10), 3)
}
