% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_summary.R
\name{stat_summary}
\alias{stat_summary}
\title{Create a stats summary}
\usage{
stat_summary(population, samples, parameter)
}
\arguments{
\item{population}{The virtual population}

\item{samples}{The drawed samples}

\item{parameter}{The parameter(s) of interest}
}
\value{
a tibble summary stats
}
\description{
This function creates a summary stats for population, samples and parameter(s) of interest.
}
\examples{
stat_summary(dplyr::tibble(x = 1:3), dplyr::tibble(x = 1:3), 'mean')

}
